fundingType
Can a fundingProject have a fundingType element?
the element is in the full metadata records for projects; to facilitate
the harvesting process, we have inserted as optional elements in the
fundingProject the following elements:
fundingType (but keep in mind that the values are prefixed with the
namespace for the ontology, i.e.
'http://w3id.org/meta-share/meta-share/', e.g.
'http://w3id.org/meta-share/meta-share/euFunds' )
funder: you can use this element for mapping the funding organization