i18n:text searialization
possible to generate `<i18n:text key="{key}">{text}</i18n:text>` if key
is not found then text is used, if text is empty it uses untranslated
(defined in sitemaps).

with this the license agreement displays either the license text (or
whatever is actually in the messages) or behaves as before.