Fix for (some) pdf reports not being send
With the default config the `lr.email.burst` is 20.  If clarin-dspace needs to send
more than `lr.email.burst` mails, some of them get queued.
The piwik-report-generator is a short lived process, it does not wait for the queue
to be empty and exits. Hence any email in the queue is actually not sent.

This fix just changes `lr.email.burst` when calling `make generate_piwik_reports`.
System properties have a rather low priority. If you have
`lr.email.burst` in `config/modules/*.cfg` or `config/dspace.cfg` this
fix will not work.